requirements.txt
# ==================================================
# TextAI Studio - Production Requirements
# ==================================================
# Generated: 2024-12-29
# Python: 3.10+
# ==================================================

# ==================================================
# CORE FRAMEWORK
# ==================================================
streamlit==1.29.0              # Web framework
python-dotenv==1.0.0           # Environment variables

# ==================================================
# MACHINE LEARNING & NLP
# ==================================================
transformers==4.36.0           # HuggingFace transformers
torch==2.1.0                   # PyTorch (CPU version)
sentence-transformers==2.2.2   # Sentence embeddings
datasets==2.15.0               # Dataset utilities

# ==================================================
# DATA PROCESSING
# ==================================================
pandas==2.1.4                  # Data manipulation
numpy==1.26.2                  # Numerical computing

# ==================================================
# VISUALIZATION
# ==================================================
plotly==5.18.0                 # Interactive charts
matplotlib==3.8.2              # Static plots
seaborn==0.13.0                # Statistical visualizations

# ==================================================
# SECURITY
# ==================================================
bcrypt==4.1.2                  # Password hashing

# ==================================================
# UTILITIES
# ==================================================
requests>=2.31.0               # HTTP requests
Pillow>=10.1.0                 # Image processing
python-multipart>=0.0.6        # File uploads

# ==================================================
# OPTIONAL: Development Tools (comment out for production)
# ==================================================
# jupyter>=1.0.0               # Jupyter notebooks
# ipykernel>=6.27.0            # Jupyter kernel
# pytest>=7.4.0                # Testing framework
# black>=23.12.0               # Code formatting
# flake8>=6.1.0                # Linting

# ==================================================
# NOTES
# ==================================================
# - Use Python 3.10 or higher
# - For GPU support, replace torch with torch-cuda version
# - Update versions periodically for security patches
# - Test thoroughly after any version updates
# ==================================================
